barcode scan is added when point pay
