+++ app/src/main/java/com/dirtfy/ppp/ui/view/tablet/record/RecordDetailScreen.kt
@@ -287,21 +287,6 @@ class RecordDetailScreen @Inject constructor(
     }
 
     @Composable
-    fun BarcodeIcon(
-        onClick: () -> Unit
-    ) {
-        val barcodeIcon = Icons.Filled.Menu
-        Icon(
-            imageVector = barcodeIcon, contentDescription = barcodeIcon.name,
-            modifier = Modifier
-                .rotate(90f)
-                .clickable {
-                    onClick()
-                }
-        )
-    }
-
-    @Composable
     fun PointDataInput(
         pointUse: UiPointUse,
         onAccountNumberChange: (String) -> Unit,
@@ -316,7 +301,7 @@ class RecordDetailScreen @Inject constructor(
                 singleLine = true,
                 keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.NumberPassword),
                 trailingIcon = {
-                    BarcodeIcon(onClick = onScanClick)
+                    Component.BarcodeIcon(onClick = onScanClick)
                 }
             )
             TextField(
